Move filters to internal ns
